address more review notes
